requirements.txt
amqp==5.1.1
arrow==1.3.0
asgiref==3.5.2
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
billiard==4.1.0
binaryornot==0.4.4
boto3==1.26.41
botocore==1.29.41
carehare==1.0.2
celery==5.3.4
celery-amqp-backend==1.1.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-rabbitmq==4.0.1
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==37.0.4
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20230430
Django==4.0.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-channels==0.7.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-import-export==3.3.5
django-ratelimit==4.1.0
django-redis==5.2.0
django-rest-authtoken==2.1.4
django-storages==1.13.2
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.7
environ==1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
Faker==19.13.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
inflection==0.5.1
ipython==8.1.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.7
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
pamqp==3.2.1
parso==0.8.3
paytmchecksum==1.7.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.3
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.16.0
Pygments==2.17.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypiwin32==223
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.7
python-nginx==1.5.7
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2022.2.1
pywatch==0.4
pywin32==306
PyYAML==6.0.1
razorpay==1.3.0
redis==5.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rich==13.7.1
s3transfer==0.6.0
service-identity==23.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stack-data==0.6.3
tablib==3.5.0
text-unidecode==1.3
traitlets==5.13.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.9
websockets==12.0
win10toast==0.9
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.1
